# src/main/resources/openapi/task-manager.yml
openapi: 3.0.3
info:
  title: Task Manager
  description: |-
    This is a Task Manager Server based on the OpenAPI 3.0 specification.
  contact:
    email: 125bharrio@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: task-manager/
tags:
  - name: Authentication
    description: Endpoints to manage authentication.
  - name: Task Management
    description: Main endpoints for task management.
  - name: User Management
    description: Endpoints to manage users.
  - name: Notifications
    description: Endpoints to manage notifications.
security:
  - BearerAuth: []
paths:
  "/api/auth/register":
    post:
      tags:
        - Authentication
      summary: User registration.
      description: "User registration. Password requirements:\n\n- Must be between 12 and 20 characters long.\n\n- Must contain at least one lowercase letter (a-z).\n\n- Must contain at least one uppercase letter (A-Z).\n\n- Must contain at least one digit (0-9).\n\n- Must contain at least one special character (@$!%*?&).\n\n- Only letters (A-Za-z), digits (0-9), and special characters (@$!%*?&) are allowed."
      operationId: register
      requestBody:
        $ref: "#/components/requestBodies/RegisterRequest"
      responses:
        "201":
          $ref: "#/components/responses/Created"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/Timeout"  
  "/api/auth/login":
    post:
      tags:
        - Authentication
      summary: User log in.
      description: User log in.
      operationId: login
      requestBody:
        $ref: "#/components/requestBodies/LoginRequest"
      responses:
        "200":
          $ref: "#/components/responses/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/Timeout"
  "/api/auth/me":
    get:
      tags:
        - Authentication
      summary: User authenticate information.
      description: User authenticate information.
      operationId: authMe
      parameters:
        - $ref: "#/components/parameters/token"
      responses:
        "200":
          $ref: "#/components/responses/AuthMeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/Timeout"
  "/api/auth/forgot_password":
    post:
      tags:
        - Authentication
      summary: Forgot password endpoint.
      description: Forgot password endpoint.
      operationId: forgotPassword
      requestBody:
        $ref: "#/components/requestBodies/ForgotPasswordRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/Timeout"
  "/api/auth/change_password":
    post:
      tags:
        - Authentication
      summary: Change password endpoint.
      description: Change password endpoint.
      operationId: changePassword
      parameters:
        - $ref: "#/components/parameters/token"
      requestBody:
        $ref: "#/components/requestBodies/ChangePasswordRequest"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/Timeout"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    token:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
        format: JWT
      example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
  schemas:
    Error:
      type: object
      description: "Data structure containing the details for errors"
      properties:
        timestamp:
          type: "string"
          format: "date-time"
          example: "2021-03-20T10:00:00Z"
        status:
          type: "integer"
          example: 400
        error:
          type: "string"
          example: "Bad Request"
        message:
          type: "string"
          example: "Invalid parameters"
        path:
          type: "string"
          example: "/api/auth/register"
        validationErrors:
          type: "array"
          items:
            type: "object"
            properties:
              field:
                type: "string"
                example: "email"
              message:
                type: "string"
                example: "Invalid email format"
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: "Complete API Design"
        description:
          type: string
          example: "Define OpenAPI specification for Task Manager."
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED]
          example: "IN_PROGRESS"
        dueDate:
          type: string
          format: date
          example: "2024-03-20"
        assignedTo:
          $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        surname:
          type: string
          description: The surname of the user.
        email:
          type: string
          format: email
          example: "john@example.com"
    RegisterRequest:
      type: object
      required:
        - email
        - name
        - surname
        - password
      properties:
        email:
          type: string
          format: email
          description: The email of the user.
        name:
          type: string
          description: The name of the user.
        surname:
          type: string
          description: The surname of the user.
        password:
          type: string
          description: The password of the user.
          minLength: 12
          maxLength: 20
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{12,20}$"
    ForgotPasswordRequest:
      type: object
      required:
        - email
        - name
      properties:
        email:
          type: string
          format: email
          description: The email of the user.
        name:
          type: string
          description: The name of the user.
    ChangePasswordRequest:
      type: object
      required:
        - email
        - password
        - newPassword
      properties:
        email:
          type: string
          format: email
          description: The email of the user.
        password:
          type: string
          description: The password of the user.
        newPassword:
          type: string
          description: The new password of the user.
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: The email of the user.
        password:
          type: string
          description: The password of the user.
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: The JWT token to authenticate the user.
  requestBodies:
    RegisterRequest:
      description: "Data structure containing information to register the user"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RegisterRequest'
    LoginRequest:
      description: "Data structure containing information to log in the user"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LoginRequest'
    ForgotPasswordRequest:
      description: "Data structure containing information to reset the user password"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ForgotPasswordRequest'
    ChangePasswordRequest:
      description: "Data structure containing information to change the user password"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ChangePasswordRequest'
  responses:
    AuthMeResponse:
      description: "User authenticated successfully"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/User"
    LoginResponse:
      description: "User logged in successfully"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/LoginResponse"
    RegisterResponse:
      description: "User registered successfully"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/User"
    Created:
      description: "Created"
    NoContent:
      description: "No content"
    BadRequest:
      description: "Bad Request"
      content:
       application/json:
         schema:
           $ref: "#/components/schemas/Error"
    Forbidden:
      description: "Forbidden"
      content:
       application/json:
         schema:
           $ref: "#/components/schemas/Error"
    Timeout:
      description: "Timeout"
      content:
       application/json:
         schema:
           $ref: "#/components/schemas/Error"
    InternalServerError:
      description: "Internal Server Error"
      content:
       application/json:
         schema:
           $ref: "#/components/schemas/Error"
    NotFound:
      description: "Not Found"
      content:
       application/json:
         schema:
           $ref: "#/components/schemas/Error"
    UnsupportedMediaType:
      description: "Unsupported Media Type"
      content:
       application/json:
         schema:
           $ref: "#/components/schemas/Error"
    ServiceUnavailable:
      description: "Service Unavailable"
      content:
       application/json:
         schema:
           $ref: "#/components/schemas/Error"
    Unauthorized:
      description: "Unauthorized"
      content:
       application/json:
         schema:
           $ref: "#/components/schemas/Error"